refactor:domain layer package 정리
