+++ app/src/main/java/com/codesoom/assignment/domain/user/User.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.domain;
+package com.codesoom.assignment.domain.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
